播放器增加切换声道功能 增加番剧搜索功能 首页全新改版

+++ DanDanPlayForiOS/ViewController/我的/View/DDPSettingDownloadTableViewCell.m
@@ -38,7 +38,7 @@
         self.countLabel.text = @"99+";
     }
     else {
-        self.countLabel.text = [NSString stringWithFormat:@"%ld", _downLoadCount];
+        self.countLabel.text = [NSString stringWithFormat:@"%lu", (unsigned long)_downLoadCount];
     }
 }
